build: Initialisiere Projekt-Setup und Code-Qualitätstools
- Fügt requirements.txt für die Projekt-Abhängigkeiten hinzu.
- Konfiguriert black, isort, flake8 und pylint über pyproject.toml und .flake8.
- Formatiert den bestehenden Code gemäß den neuen Richtlinien.
- Stellt eine saubere Code-Basis für zukünftige Entwicklungen her.
