+++ requirements.txt
@@ -1,0 +1,2 @@
+ortools>=9.5
+matplotlib>=3.5.0
